# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  name: string;
  icon: string;
  serviceCount: number;
}

export interface Service {
  id: string;
  categoryId: string;
  name: string;
  description: string;
  imageUrl: string;
  startingPrice: number;
  duration: string;
  rating: number;
  bookings: number;
  accent: string;
}

export interface AddOn {
  id: string;
  name: string;
  price: number;
}

export interface Faq {
  question: string;
  answer: string;
}

export interface Review {
  id: string;
  customerName: string;
  rating: number;
  comment: string;
  date: string;
}

export type ServiceDetail = Service & {
  included: string[];
  excluded: string[];
  addOns: AddOn[];
  warranty: string;
  faqs: Faq[];
  reviews: Review[];
};

export interface Professional {
  id: string;
  name: string;
  photoUrl: string;
  verified: boolean;
  rating: number;
  totalJobs: number;
  experience: string;
  skills: string[];
  availability: string;
  startingPrice: number;
  area: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: string;
  service: Service;
  professional: Professional;
  status: BookingStatus;
  scheduledAt: string;
  address: string;
  total: number;
  eta: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
}

export type BookingInputPaymentMethod = typeof BookingInputPaymentMethod[keyof typeof BookingInputPaymentMethod];


export const BookingInputPaymentMethod = {
  upi: 'upi',
  card: 'card',
  netbanking: 'netbanking',
  wallet: 'wallet',
  cash: 'cash',
} as const;

export interface BookingInput {
  serviceId: string;
  /** @nullable */
  professionalId?: string | null;
  scheduledAt: string;
  address: string;
  /** @nullable */
  instructions?: string | null;
  addOnIds?: string[];
  paymentMethod: BookingInputPaymentMethod;
}

export interface Offer {
  id: string;
  title: string;
  subtitle: string;
  code: string;
  accent: string;
}

export interface HomeFeed {
  greeting: string;
  location: string;
  categories: Category[];
  popularServices: Service[];
  recommended: Service[];
  recentlyBooked: Booking[];
  offers: Offer[];
  verifiedProfessionals: Professional[];
}

export interface ProfessionalDashboard {
  professional: Professional;
  todayEarnings: number;
  todayJobs: number;
  pendingRequests: number;
  upcomingJobs: number;
  isOnline: boolean;
}

export interface AdminSummary {
  totalUsers: number;
  activeUsers: number;
  totalProfessionals: number;
  activeProfessionals: number;
  totalBookings: number;
  completedBookings: number;
  cancelledBookings: number;
  gmv: number;
  revenue: number;
  commission: number;
  averageOrderValue: number;
}

export type ListServicesParams = {
category?: string;
search?: string;
};

export type ListProfessionalsParams = {
serviceId?: string;
area?: string;
};

export type ListBookingsParams = {
status?: ListBookingsStatus;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;