# lib/api-zod/src/generated/types/addOn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AddOn {
  id: string;
  name: string;
  price: number;
}