# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { Professional } from './professional';
import type { Service } from './service';

export interface Booking {
  id: string;
  service: Service;
  professional: Professional;
  status: BookingStatus;
  scheduledAt: string;
  address: string;
  total: number;
  eta: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
}