# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputPaymentMethod } from './bookingInputPaymentMethod';

export interface BookingInput {
  serviceId: string;
  /** @nullable */
  professionalId?: string | null;
  scheduledAt: string;
  address: string;
  /** @nullable */
  instructions?: string | null;
  addOnIds?: string[];
  paymentMethod: BookingInputPaymentMethod;
}