# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: string;
  title: string;
  subtitle: string;
  code: string;
  accent: string;
}