# lib/api-zod/src/generated/types/professionalDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Professional } from './professional';

export interface ProfessionalDashboard {
  professional: Professional;
  todayEarnings: number;
  todayJobs: number;
  pendingRequests: number;
  upcomingJobs: number;
  isOnline: boolean;
}