# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  customerName: string;
  rating: number;
  comment: string;
  date: string;
}