# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEA home-services marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  categoryId: string;
  name: string;
  description: string;
  imageUrl: string;
  startingPrice: number;
  duration: string;
  rating: number;
  bookings: number;
  accent: string;
}