# lib/db/src/schema/reviews.ts
import { pgTable, text, numeric, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const reviewsTable = pgTable("sea_reviews", {
  id: text("id").primaryKey(),
  serviceId: text("service_id").notNull(),
  customerName: text("customer_name").notNull(),
  rating: numeric("rating", { precision: 3, scale: 2 }).notNull(),
  comment: text("comment").notNull(),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertReviewSchema = createInsertSchema(reviewsTable).omit({ createdAt: true });
export type InsertReview = z.infer<typeof insertReviewSchema>;
export type Review = typeof reviewsTable.$inferSelect;